Move: moving get/post/delete Profile to userService.ts :: lib\api\authService.ts
