+++ src/lib/api/userService.ts
@@ -1,12 +1,45 @@
 import { QueryFunctionContext } from "@tanstack/react-query";
-import instance from "./instance";
+import { Profile, ProfileData, ProfileEdittable, User } from "@/types/types";
+import { get, patch, post, remove } from "./method";
+
+export async function getUserInfo(userId: string): Promise<ProfileData> {
+  try {
+    const res = await get(`/profile/${userId}`);
+    return res.data;
+  } catch (err) {
+    throw err;
+  }
+}
+
+export async function postProfile(data: Profile): Promise<ProfileData> {
+  try {
+    const res = await post<Profile>("/profile", data);
+    return res.data;
+  } catch (err) {
+    throw err;
+  }
+}
+
+export async function patchProfile(
+  userId: string,
+  data: Partial<ProfileEdittable>,
+): Promise<{
+  user: User;
+}> {
+  try {
+    const res = await patch<Partial<ProfileEdittable>>(`/profile/${userId}`, data);
+    return res.data;
+  } catch (err) {
+    throw err;
+  }
+}
 
 export const getFavoriteTrainers = async ({
   pageParam = 1,
   queryKey,
 }: QueryFunctionContext<[string, { userId: string }]>) => {
   const [, { userId }] = queryKey;
-  const { data } = await instance.get(`/users/${userId}/favorites`, {
+  const { data } = await get(`/users/${userId}/favorites`, {
     params: { page: pageParam },
   });
   return data;
@@ -19,7 +52,7 @@ export const getFavorite = async (
   favoriteTotalCount: number;
 }> => {
   try {
-    const res = await instance.get(`/trainers/${trainerId}/favorite`);
+    const res = await get(`/trainers/${trainerId}/favorite`);
     return res.data;
   } catch (err) {
     throw err;
@@ -35,9 +68,9 @@ export const toggleFavorite = async (
 }> => {
   try {
     if (isFavorite) {
-      return (await instance.delete(`/trainers/${trainerId}/favorite`)).data;
+      return (await remove(`/trainers/${trainerId}/favorite`)).data;
     }
-    return (await instance.post(`/trainers/${trainerId}/favorite`)).data;
+    return (await post<undefined>(`/trainers/${trainerId}/favorite`, undefined)).data;
   } catch (err) {
     throw err;
   }
